im so tired...
Đã thực hiện các thay đổi lớn trong mã nguồn, bao gồm:
- Sử dụng `std::filesystem` cho quản lý đường dẫn tệp (cải thiện tính an toàn và chức năng).
- Thêm hằng số mới cho các phần mở rộng tệp liên quan đến osu! (tăng cường tính rõ ràng và khả năng bảo trì).
- Tổ chức lại không gian tên và mã nguồn (cải thiện cấu trúc mã).
- Chuyển đổi nhiều hàm sang sử dụng `std::filesystem` (tăng cường độ tin cậy khi thao tác với tệp).
- Giới thiệu các lớp và phương thức mới để xử lý âm thanh và skin (quản lý loại tệp hỗ trợ hiệu quả hơn).
- Cải tiến xử lý đối tượng hit trong logic xử lý beatmap (quản lý bộ nhớ tốt hơn với con trỏ thông minh).
- Thêm xử lý lỗi và cải tiến ghi log với `spdlog` (nâng cao khả năng gỡ lỗi).
- Cập nhật phong cách mã cho tính nhất quán (sử dụng các tính năng C++ hiện đại).
- Cập nhật tệp dự án để phản ánh cấu trúc mới (bao gồm các tệp nguồn và tiêu đề mới).

+++ include/structures/render/collection.h
@@ -12,7 +12,7 @@ namespace Structures::Render::Objects
 
 	public:
 		std::vector<std::variant<ObjectShared, PolyObjectShared>> data;
-		std::vector<std::pair<size_t, size_t>> render_range; // Phần tử đầu tiên là 0
+		std::vector<std::pair<int64_t, int64_t>> render_range; // Phần tử đầu tiên là 0
 		SDL_FPoint offset = { 0, 0 };
 		bool visible = true;
 		void render(const SDL_FPoint& camera_offset) const;
